Cambios en las entidades y en producto, implementacion de la funcion search y aplicacion de descuentos

+++ ecommerce-back/src/main/java/com/ecommerce/dto/Discount/DiscountRuleAdminDTO.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -15,7 +16,7 @@ public class DiscountRuleAdminDTO {
     private Long id;
     private LocalDate startDate;
     private LocalDate endDate;
-    private Double percentage;
+    private BigDecimal percentage;
     private Boolean state;
     private Long categoryId;
     private Boolean deleted;
