+++ ecommerce-back/src/main/java/com/ecommerce/dto/Discount/UpdateDiscountRuleDTO.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -15,7 +16,7 @@ public class UpdateDiscountRuleDTO {
     private Boolean deleted;
     private LocalDate startDate;
     private LocalDate endDate;
-    private Double percentage;
+    private BigDecimal percentage;
     private Boolean state;
     private Long categoryId;
     private Long productId;
